Refactoring of CLI test case calls
